requirements.txt
pandas
duckdb
PyYAML
requests